refactor to implement dependency injection at all layers
